Fix alert names in values.yaml
For some of the alerts we have names that don't match the current name of the alert in kube-prometheus-stack.

See, e.g. https://github.com/prometheus-community/helm-charts/search?q=KubeDaemonsetMisscheduled

This adds on the correct names while leaving the old ones as we originally added them under that name for specific users.